Moved ProcessStep-containers to shared/containers
